11.7.1 Правка бага: Тип значения delta таблицы metrics не соответствовал типа delta dto.Metrics.

+++ migration/postgres/20241207160536_create_db.up.sql
@@ -30,7 +30,7 @@ VALUES ('counter');
 CREATE TABLE IF NOT EXISTS metrics
 (
     ID             VARCHAR(100) UNIQUE,
-    delta          INTEGER          NULL,
+    delta BIGINT NULL,
     value          DOUBLE PRECISION NULL,
     metric_type_id SMALLINT,
     PRIMARY KEY (ID),
